% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finddxtemplate-package.R
\docType{package}
\name{finddxtemplate-package}
\alias{finddxtemplate}
\alias{finddxtemplate-package}
\title{finddxtemplate: Tools to create a Rmd template for html documents}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Tools to create a Rmd template for html documents.
}
\details{
Open the pkgdown : \code{finddxtemplate::open_pkgdown()}
}
\author{
\strong{Maintainer}: Arthur Bréant \email{arthur@thinkr.fr} (\href{https://orcid.org/0000-0003-1668-0963}{ORCID})

Authors:
\itemize{
  \item Margot Brard \email{margot@thinkr.fr} (\href{https://orcid.org/0000-0001-6754-0659}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
