% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_document_find.R
\name{html_document_find}
\alias{html_document_find}
\title{Convert a .Rmd file into a .html file according to finddx design / template.}
\usage{
html_document_find(
  toc = TRUE,
  toc_depth = 4,
  toc_float = TRUE,
  code_folding = "show",
  ...
)
}
\arguments{
\item{toc}{Boolean. To include a table of contents in the output. Default is TRUE.}

\item{toc_depth}{Numeric. Depth of headers to include in table of contents. Default is 4.}

\item{toc_float}{Boolean. To float the table of contents to the left of the document. Default is TRUE.}

\item{code_folding}{Character. Enable document readers to toggle the display of R code chunks.}

\item{...}{Additional arguments of \code{rmarkdown::\link{html_document}()}.}
}
\value{
An R Markdown output format object
}
\description{
Convert a .Rmd file into a .html file according to finddx design / template.
}
\examples{
# Store in temp file for the example
rmddir <- tempfile(pattern = "rmd-")
dir.create(rmddir)

# Knit
rmarkdown::render(input = system.file("rmarkdown", "rmd_template.Rmd", package = "finddxtemplate"), 
                  output_format = html_document_find(),
                  output_dir = rmddir)

# Open the knitted example
browseURL(file.path(rmddir, "rmd_template.html"))
}
