% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_document_finddx.R
\name{html_document_finddx}
\alias{html_document_finddx}
\title{Convert a .Rmd file into a .html file according to finddx design / template.}
\usage{
html_document_finddx(
  toc = TRUE,
  toc_depth = 4,
  toc_float = TRUE,
  code_folding = "show",
  ...
)
}
\arguments{
\item{toc}{Boolean. To include a table of contents in the output. Default is TRUE.}

\item{toc_depth}{Numeric. Depth of headers to include in table of contents. Default is 4.}

\item{toc_float}{Boolean. To float the table of contents to the left of the document. Default is TRUE.}

\item{code_folding}{Character. Enable document readers to toggle the display of R code chunks.}

\item{...}{Additional arguments of \code{rmarkdown::\link{html_document}()}.}
}
\value{
An R Markdown output format object
}
\description{
Convert a .Rmd file into a .html file according to finddx design / template.
}
\examples{
rmarkdown::render(
  input = system.file("rmarkdown", "template.Rmd", package = "finddxtemplate"), 
  output_format = html_document_finddx()
)
}
